Rework library initialization.
The `Library` class keeps books in a dictionary keyed on the path of the
book.  Renamed `-scanDirsForBooks` to `-startScanningForBooks` and run
the scan in a global background queue, sending a notification when done.

In the app delegate, update the sorted copy of the books when finished
scanning rather than updating the library instance variable.