Use an `NSOrderedSet` in the `Library` class.
Change the `books` property in `Library` from a dictionary of paths to
books into an `NSMutableOrderedSet` of books internally with a read only
property of type `NSOrderedSet`.  Add a `-sortBy:` method to `Library`
to sort the `_books` instance variable.

Remove the copied `_books` instance variable from the app delegate and
use the `_library` instance variable directly as the backing data for
the table.

In `Logger`, add the index and count of books to the log message for
book selection.