Added minizip-ng 4.0.7 for reading EPUB files.
Vendored the source distribution of [minizip-ng][1] into the `libs/`
directory and added a Framework target for it.  Also created umbrella
header `minizip.h` for the framework.

Extended the `EPUB` class init method to open the EPUB zip file and look
for the `container.xml` file in the manifest and as a fall-back, find
any `.opf` file.

[1]: https://github.com/zlib-ng/minizip-ng