Extend `EPUB` class to find the package XML file.
Add code to `-initWithPath:` that reads from the EPUB zip file and
attempts to locate the `container.xml` file, parse it and extract the
first "rootfile" that points to an OPF package XML file.  Fall back to
scanning the zip archive for files ending in `.opf` if the container
isn't found, isn't readable or is missing an OPF package rootfile.

Still TODO: parse the OPF package file to extract metadata.

Added the `EPUBContainer` class that parses the `container.xml` file,
and the `EPUBRootfile` class that represents a `<rootfile>` tag found in
the container file.

Added "copy" to some properties where it was missing, and marked some
init methods with `NS_UNAVAILABLE` where needed.