Add function `formatForExtension()`.
Rename `formatName()` to `nameForFormat()` to follow the pattern of
other format helpers.  Internally, add static function
`getformatProperties()` to do one-time initialization of the format
properties.

Change `Book` `-initWithPath:andFileSize:` method to use
`formatForExtension()` to eliminate the ladder of if statements.  And
since we're storing the `Class<File>` on the `Book`, get the Format enum
value from the file class instead of setting it as a property on `Book`.