Make `OPFIdentifier` immutable.
Also drop the `ID` property from `OPFIdentifier` since it's really part
of the underlying XML document, not the identifier itself.

Change `OPFPackage` to hold onto the attributes dictionary when the
`<dc:identifier>` element starts and use it on element end to create the
`OPFIdentifier` rather than create it at element start and finish
initializing it at element end when the text content is available.